fix: corrige migration

+++ database/migrations/2024_11_27_222438_add_columns_in_candidate_attachments_table.php
@@ -1,5 +1,4 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -12,10 +11,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('candidate_attachments', function (Blueprint $table) {
-            $table->unsignedBigInteger('selection_process_id')->nullable()->after('candidate_id');
-            $table->foreign('selection_process_id')
-                ->references('id')
-                ->on('selection_process');
+            if (!Schema::hasColumn('candidate_attachments', 'selection_process_id')) {
+                $table->unsignedBigInteger('selection_process_id')->nullable()->after('candidate_id');
+                $table->foreign('selection_process_id')
+                    ->references('id')
+                    ->on('selection_process');
+            }
         });
     }
 
@@ -25,8 +26,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('candidate_attachments', function (Blueprint $table) {
-            $table->dropForeign(['selection_process_id']);
-            $table->dropColumn(['selection_process_id']);
+            if (Schema::hasColumn('candidate_attachments', 'selection_process_id')) {
+                $table->dropForeign(['selection_process_id']);
+                $table->dropColumn(['selection_process_id']);
+            }
         });
     }
 };
